srm: Drop fail fast for protocol checks
For both get and put the requested protocol is compared to supported
protocols before creating the schedulable request objects. I guess
this was supposed to provide a fail-fast path. However it means we

- Don't leave behind the request object for subsequent analysis.

- Do a non-trivial operation in the Jetty thread when we have
  a perfectly fine scheduler to deal with this.

- Duplicate checks as the Storage class already checks this while
  selecting a door.

This patch gets rids of the duplication. I submit this for 2.9 as is
resolves a regression (in the old code the protocols were fetched
differently).

Target: trunk
Request: 2.9
Require-notes: yes
Require-book: no
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Patch: http://rb.dcache.org/r/6977/