Drop messages when TTL is exceeded
The purpose of having a message TTL is to allow messages to be
discarded when the TTL is exceeded. We have used this sparingly
in a couple of services, but there is no reason why we cannot
apply this logic to all services.

We have often observed that dCache doesn't recover well from
an overload situation and I hope this change improves on that.

Target: trunk
Require-notes: yes
Require-book: no
Acked-by: Paul Millar <paul.millar@desy.de>
Patch: http://rb.dcache.org/r/6994/