xrootd: Quote IPv6 addresses in kXR_redirect replies
The client concatenates the returned host name and the
binary encoded port number into a URI. This fails if
the host name is an IPv6 address as the client fails to
quote the address.

The xrootd spec doesn't say that the server should quote
the address, however the answer we got from the xrootd
developers is that all IPv6 addresses must be quoted by
the server.

This patch quotes the address in the reply. Until the spec
is updated, I consider this a workaround.

Target: trunk
Require-notes: yes
Require-book: no
Request: 2.9
Request: 2.8
Request: 2.7
Request: 2.6
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Acked-by: Paul Millar <paul.millar@desy.de>
Patch: https://rb.dcache.org/r/7046/