webdav: allow users to choose between showing file's content inline and downloading it
This patch adjusts the directory listing to be more "browser friendly".
There are two parts to this: updating the GET operation and updating the template.

When a pool processes a GET or HEAD request, it generates a Content-Disposition header.  This
is intended to provide the client with a hint under which filename the file should be saved,
but has a side-effect of forcing the pool to suggest either showing the content in the browser
('inline') or saved as a file ('attachment').  The current behaviour is to always sent
'attachment'.

There are situations where a user may wish to view the content in the browser (assuming the
browser supports viewing this content-type) that this behaviour precludes.

With this patch, an optional query parameter 'action' may be specified.  If the action
parameter has the value 'show' then dCache will return a Content-Disposition of 'inline'; if
action is 'download' then the a Content-Disposition is 'attachment'.  If the action parameter
isn't one of these two, or no action parameter is specied then 'attachment' is used.

This patch also updates the html rendering of a directory with the following new behaviour:

  selecting the filename will deliver the file's contents with Content-Disposition
  of 'inline' (i.e. "viewing the file")

  an additional 'download' icon is shown next to each file's filename; selecting
  this will deliver the file's contents with Content-Disposition of 'attachment'.
  This also includes the HTML5 'download' attribute that provides the same hint
  from HTML.

  both the filename and the download icon now have titles (rendered as tooltips in
  most browsers) that describe the effect of selecting the links.

Target: master
Patch: https://rb.dcache.org/r/7058/
Acked-by: Gerd Behrmann
Require-notes: yes
Require-book: yes