dcache alarms: add properties and fuller documentation for configuring SMTP appender
The instructions in the current logback-server.xml are
incomplete in terms of how to set up the appender to
talk to a normal SMTP server.

This patch is a revision of http://rb.dcache.org/r/6945
which exposes all of the configuration through the
standard dCache property framework.

This means that the patch is more invasive,
with code slightly modified in a few places.

A separate patch adding the fuller settings to
logback-server.xml without property changes
has already been posted for backport and merged
in (2.6, 2.7, 2.8 and 2.9).

Testing: Used new properties to configure send through gmail SMTP
to two fnal email recipients.

Target: master
Require-book: yes
Require-notes: yes
Acked-by: Karsten

RELEASE NOTES:  Fuller documentation for the configuration
of the alarm server's logback appenders (particularly for
email notification) has been provided.  The logback
settings are now fully configurable through the properties
framework.