ftp: add support for command pipelining
Motivation:

The Globus transfer service issues a RETR command while and existing
RETR transfer is underway.  It can also issue a CKSM command while a
corresponding STOR command is underway.

RFC 959 describes how "certain commands" may be issued while a transfer
is underway, suggesting that there are some commands that may not be
issued under these circumstances.  The language is ambigious about
whether the provided list (ABOR, STAT and QUIT) are examples of allowed
commands or the complete set. Moreover, no RFC document describes how
the server should react when a transfer is underway and the client
issues commands that are not allowed during a transfer.

Current dCache behaviour is to fail such commands immediately.  This
leads to an inversion of the responses; for example, a client issuing a
pipeline of two RETR commands may receive responses from dCache like:

    150 Opening BINARY data connection for /public/test-1
    503 Transfer in progress
    226 Transfer complete.

with the 503 response is from dCache failing the second RETR.

The Globus transfer service interprets the 5xx response as the first
RETR command having failed.  This results in Globus transfer service
tearing down all control channels (aborting all ongoing transfers),
waiting 5 minutes and retrying.

In contrast, Globus transfer service server responds to pipeline
commands by processing them in order.  For example, a pipeline of RETR,
DELE commands results in the file being deleted only after the transfer
completes.

Similar to dCache, the Globus transfer service server reacts to ABOR
requests immediately.  However, the behaviour is not spec compliant: RFC
959 states:

    4.1.3. FTP Service Commands; ABORT (ABOR)

    The control connection is not to be closed by the server [...]

yet a pipeline of RETR, ABOR, DELE commands results in the transfer
being aborted (after some tiny fragment of the file was transmitted) and
the server closing the control channel without deleting the file.  The
responses before the control channel was closed are:

    150 Beginning transfer.
    226 Transfer Complete.

The ABOR is processed immediate, irrespective of what other commands are
pending.  For example, a pipeline of RETR, RETR, ABOR, DELE commands
results in the first transfer being aborted and the server closing the
control channel without deleting the file.  The responses is as before:

    150 Beginning transfer.
    226 Transfer Complete.

When an ABOR command is issued with no transfer underway (to which the
server correctly responds with a 226 reply) and then a pipeline of RETR,
ABOR, DELE commands is issued then the transfer is aborted (after some
tiny fragment of the file was transmitted) and the server will close the
control channel without deleting the file.  However, the responses
before the control channel is closed are now:

    150 Beginning transfer.
    500-Command failed. : callback failed.
    500-Data connection has been closed due to a call to globus_ftp_control_data_for
    500 End.

This apparently broken behaviour may explain why the Globus transfer
service does not issue ABOR commands but rather tears down the control
channel when aborting a transfer.

Modification:

Use annotation to identify commands that may run concurrently with a
transfer.  Non-concurrent requests are queued and processed sequentially
once the transfer has finished.  Concurrent requests (such as ABOR) are
processed immediately, but their replies are captured and returned as if
the command was processed in the correct sequence.

For example, if a client issues a pipeline of RETR, RETR, RETR ABOR
commands then it may receive the following responses:

    150 Opening BINARY data connection for /public/test-1
    426 Transfer aborted
    150 Opening BINARY data connection for /public/test-2
    226 Transfer complete.
    150 Opening BINARY data connection for /public/test-3
    226 Transfer complete.
    226 Abort successful

Result:

Better compatibility with Globus transfer service when transferring
multiple files from dCache.

Target: master
Request: 3.0
Request: 2.16
Requires-notes: yes
Requires-book: no
Patch: https://rb.dcache.org/r/10104/
Acked-by: Tigran Mkrtchyan