dcache-webadmin: fix pool queue plot grid placeholder image
The pool queue plot grid is computed to find the best square fit up to 8 X 8 of plots.
This means that there will remainders (blanks) for all but the square cases.
Previously the placeholder was an empty symbol denoting no pool, but this particular
image was not the right size and at any rate was rather ugly.

The patch replaces the previous image with a simple blank image the same size as the plots.

Target:  2.6
Require-notes: no
Require-book: no
Acked-by: Dmitry
Committed: 7fed8893e0e718cd053f80e6ae3f4c432d4ed9fe