srm: fix status of all-failed container jobs
The status of a container job is updated as a side-effect of running
the ContainerRequest#getRequestStatus method.  Unfortunately, if
all the file requests have failed then a very long, multi-line
status is recorded, which includes a copy of the previous
status.

This patch fixes the status so that, even under these circumstances,
the status line is a short string.

Target: master
Requires-notes: no
Requires-book: no