network utils: add constant for canonical host name
This patch is an add-on to 7127.  Since a number of alarm keys
need to include the host name in the set of identifiers,
the static initializer has been extended to make this
readily accessible as a static final field.

Target: master
Patch: https://rb.dcache.org/r/7140/
Require-notes: no
Require-book: no
Acked-by: Tigran