alarms: eliminate httpd.alarms.db.xml.path
Since it is not possible to stream an xml
file by giving the DataNucleus connector
a remote URL, the constraint that httpd
and alarms share a local (or mounted)
file partition for alarms data still
holds when the alarms.db.type is 'xml'.

The alarms.db.xml.path property is
a convenience so as not to have to define
the entire URL in case one wishes to
change the location of the store xml
file from the default.

The immutable httpd version of this
property was thus merely a mirror
defined for use in spring injection
(for internal use only).

However, it should be noted that,
subsequent to changes made in 02/2014
to the webadmin spring injection,
the data stores are now entirely
configured via spring profiles;
hence, the internal xml path field no
longer serves any purpose in
the dCache DataNucleusAlarmStore class.

This patch eliminates the (immutable)
httpd path property and also the
internal xml path field.

Testing:

Deployed, tested the various type
options on the httpd end (rdbms, xml, off).

Target: master
Patch: https://rb.dcache.org/r/7143/
Require-notes: no
Require-book: no
Acked-by: Tigran